refactor(server): replace ToolSupport with Tuist.MCP.Tool macro across all MCP tools
Migrate all MCP tools to use the Tuist.MCP.Tool macro with authorize keyword
list syntax and MCPTool alias, eliminating the ToolSupport module entirely.
