+++ server/lib/tuist/mcp/components/tools/get_bundle_artifact_tree.ex
@@ -17,6 +17,7 @@ defmodule Tuist.MCP.Components.Tools.GetBundleArtifactTree do
     }
 
   alias Tuist.Bundles
+  alias Tuist.MCP.Tool, as: MCPTool
 
   @impl EMCP.Tool
   def description,
@@ -25,7 +26,7 @@ defmodule Tuist.MCP.Components.Tools.GetBundleArtifactTree do
 
   def execute(conn, %{"bundle_id" => bundle_id}) do
     with {:ok, _, _} <-
-           ToolSupport.load_and_authorize(
+           MCPTool.load_and_authorize(
              Bundles.get_bundle(bundle_id),
              conn.assigns,
              :read,
