+++ server/lib/tuist/mcp/components/tools/get_xcode_build.ex
@@ -18,6 +18,7 @@ defmodule Tuist.MCP.Components.Tools.GetXcodeBuild do
 
   alias Tuist.Builds
   alias Tuist.MCP.Formatter
+  alias Tuist.MCP.Tool, as: MCPTool
 
   @impl EMCP.Tool
   def description,
@@ -28,7 +29,7 @@ defmodule Tuist.MCP.Components.Tools.GetXcodeBuild do
     build_run_id = Map.get(args, "build_run_id")
 
     with {:ok, build, _project} <-
-           ToolSupport.load_and_authorize(
+           MCPTool.load_and_authorize(
              get_build(build_run_id),
              conn.assigns,
              :read,
