+++ server/lib/tuist/mcp/components/tools/list_bundles.ex
@@ -5,6 +5,7 @@ defmodule Tuist.MCP.Components.Tools.ListBundles do
 
   use Tuist.MCP.Tool,
     name: "list_bundles",
+    authorize: [action: :read, category: :bundle],
     schema: %{
       "type" => "object",
       "properties" => %{
@@ -34,49 +35,47 @@ defmodule Tuist.MCP.Components.Tools.ListBundles do
 
   alias Tuist.Bundles
   alias Tuist.MCP.Formatter
+  alias Tuist.MCP.Tool, as: MCPTool
 
   @impl EMCP.Tool
   def description,
     do:
       "List bundles (app binaries) for a project. The account_handle and project_handle can be extracted from a Tuist dashboard URL: #{Tuist.Environment.app_url()}/{account_handle}/{project_handle}."
 
-  def execute(conn, args) do
-    with {:ok, project} <-
-           ToolSupport.resolve_and_authorize_project(args, conn.assigns, :read, :bundle) do
-      page = ToolSupport.page(args)
-      page_size = ToolSupport.page_size(args)
-      filters = build_filters(project.id, args)
+  def execute(_conn, args, project) do
+    page = MCPTool.page(args)
+    page_size = MCPTool.page_size(args)
+    filters = build_filters(project.id, args)
 
-      {bundles, meta} =
-        Bundles.list_bundles(%{
-          filters: filters,
-          order_by: [:inserted_at],
-          order_directions: [:desc],
-          page: page,
-          page_size: page_size
-        })
+    {bundles, meta} =
+      Bundles.list_bundles(%{
+        filters: filters,
+        order_by: [:inserted_at],
+        order_directions: [:desc],
+        page: page,
+        page_size: page_size
+      })
 
-      {:ok,
-       %{
-         bundles:
-           Enum.map(bundles, fn bundle ->
-             %{
-               id: bundle.id,
-               name: bundle.name,
-               app_bundle_id: bundle.app_bundle_id,
-               version: bundle.version,
-               type: to_string(bundle.type),
-               supported_platforms: bundle.supported_platforms || [],
-               install_size: bundle.install_size,
-               download_size: bundle.download_size,
-               git_branch: bundle.git_branch,
-               git_commit_sha: bundle.git_commit_sha,
-               inserted_at: Formatter.iso8601(bundle.inserted_at)
-             }
-           end),
-         pagination_metadata: ToolSupport.pagination_metadata(meta)
-       }}
-    end
+    {:ok,
+     %{
+       bundles:
+         Enum.map(bundles, fn bundle ->
+           %{
+             id: bundle.id,
+             name: bundle.name,
+             app_bundle_id: bundle.app_bundle_id,
+             version: bundle.version,
+             type: to_string(bundle.type),
+             supported_platforms: bundle.supported_platforms || [],
+             install_size: bundle.install_size,
+             download_size: bundle.download_size,
+             git_branch: bundle.git_branch,
+             git_commit_sha: bundle.git_commit_sha,
+             inserted_at: Formatter.iso8601(bundle.inserted_at)
+           }
+         end),
+       pagination_metadata: MCPTool.pagination_metadata(meta)
+     }}
   end
 
   defp build_filters(project_id, args) do
