+++ server/lib/tuist/mcp/components/tools/list_generations.ex
@@ -5,6 +5,7 @@ defmodule Tuist.MCP.Components.Tools.ListGenerations do
 
   use Tuist.MCP.Tool,
     name: "list_generations",
+    authorize: [action: :read, category: :run],
     schema: %{
       "type" => "object",
       "properties" => %{
@@ -38,53 +39,51 @@ defmodule Tuist.MCP.Components.Tools.ListGenerations do
 
   alias Tuist.CommandEvents
   alias Tuist.MCP.Formatter
+  alias Tuist.MCP.Tool, as: MCPTool
 
   @impl EMCP.Tool
   def description,
     do:
       "List generation runs for a project. The account_handle and project_handle can be extracted from a Tuist dashboard URL: #{Tuist.Environment.app_url()}/{account_handle}/{project_handle}."
 
-  def execute(conn, args) do
-    with {:ok, project} <-
-           ToolSupport.resolve_and_authorize_project(args, conn.assigns, :read, :run) do
-      page = ToolSupport.page(args)
-      page_size = ToolSupport.page_size(args)
-      filters = build_filters(project.id, args)
+  def execute(_conn, args, project) do
+    page = MCPTool.page(args)
+    page_size = MCPTool.page_size(args)
+    filters = build_filters(project.id, args)
 
-      {events, meta} =
-        CommandEvents.list_command_events(%{
-          filters: filters,
-          order_by: [:ran_at],
-          order_directions: [:desc],
-          page: page,
-          page_size: page_size
-        })
+    {events, meta} =
+      CommandEvents.list_command_events(%{
+        filters: filters,
+        order_by: [:ran_at],
+        order_directions: [:desc],
+        page: page,
+        page_size: page_size
+      })
 
-      {:ok,
-       %{
-         generations:
-           Enum.map(events, fn event ->
-             %{
-               id: event.id,
-               duration: event.duration,
-               status: status_to_string(event.status),
-               tuist_version: event.tuist_version,
-               swift_version: event.swift_version,
-               macos_version: event.macos_version,
-               is_ci: event.is_ci,
-               git_branch: event.git_branch,
-               git_commit_sha: event.git_commit_sha,
-               git_ref: event.git_ref,
-               command_arguments: event.command_arguments,
-               cacheable_targets: event.cacheable_targets,
-               local_cache_target_hits: event.local_cache_target_hits,
-               remote_cache_target_hits: event.remote_cache_target_hits,
-               ran_at: Formatter.iso8601(event.created_at, naive: :utc)
-             }
-           end),
-         pagination_metadata: ToolSupport.pagination_metadata(meta)
-       }}
-    end
+    {:ok,
+     %{
+       generations:
+         Enum.map(events, fn event ->
+           %{
+             id: event.id,
+             duration: event.duration,
+             status: status_to_string(event.status),
+             tuist_version: event.tuist_version,
+             swift_version: event.swift_version,
+             macos_version: event.macos_version,
+             is_ci: event.is_ci,
+             git_branch: event.git_branch,
+             git_commit_sha: event.git_commit_sha,
+             git_ref: event.git_ref,
+             command_arguments: event.command_arguments,
+             cacheable_targets: event.cacheable_targets,
+             local_cache_target_hits: event.local_cache_target_hits,
+             remote_cache_target_hits: event.remote_cache_target_hits,
+             ran_at: Formatter.iso8601(event.created_at, naive: :utc)
+           }
+         end),
+       pagination_metadata: MCPTool.pagination_metadata(meta)
+     }}
   end
 
   defp build_filters(project_id, args) do
