+++ server/lib/tuist/mcp/components/tools/list_test_cases.ex
@@ -5,6 +5,7 @@ defmodule Tuist.MCP.Components.Tools.ListTestCases do
 
   use Tuist.MCP.Tool,
     name: "list_test_cases",
+    authorize: [action: :read, category: :test],
     schema: %{
       "type" => "object",
       "properties" => %{
@@ -49,6 +50,7 @@ defmodule Tuist.MCP.Components.Tools.ListTestCases do
     }
 
   alias Tuist.MCP.Formatter
+  alias Tuist.MCP.Tool, as: MCPTool
   alias Tuist.Tests
 
   @impl EMCP.Tool
@@ -56,42 +58,39 @@ defmodule Tuist.MCP.Components.Tools.ListTestCases do
     do:
       "List test cases for a project. The account_handle and project_handle can be extracted from a Tuist dashboard URL: #{Tuist.Environment.app_url()}/{account_handle}/{project_handle}."
 
-  def execute(conn, args) do
-    with {:ok, project} <-
-           ToolSupport.resolve_and_authorize_project(args, conn.assigns, :read, :test) do
-      page = ToolSupport.page(args)
-      page_size = ToolSupport.page_size(args)
-      filters = build_filters(args)
+  def execute(_conn, args, project) do
+    page = MCPTool.page(args)
+    page_size = MCPTool.page_size(args)
+    filters = build_filters(args)
 
-      {test_cases, meta} =
-        Tests.list_test_cases(project.id, %{
-          filters: filters,
-          order_by: [:last_ran_at],
-          order_directions: [:desc],
-          page: page,
-          page_size: page_size
-        })
+    {test_cases, meta} =
+      Tests.list_test_cases(project.id, %{
+        filters: filters,
+        order_by: [:last_ran_at],
+        order_directions: [:desc],
+        page: page,
+        page_size: page_size
+      })
 
-      {:ok,
-       %{
-         test_cases:
-           Enum.map(test_cases, fn test_case ->
-             %{
-               id: test_case.id,
-               name: test_case.name,
-               module_name: test_case.module_name,
-               suite_name: test_case.suite_name,
-               is_flaky: test_case.is_flaky,
-               is_quarantined: test_case.is_quarantined,
-               last_status: to_string(test_case.last_status),
-               last_duration: test_case.last_duration,
-               last_ran_at: Formatter.iso8601(test_case.last_ran_at, naive: :utc),
-               avg_duration: test_case.avg_duration
-             }
-           end),
-         pagination_metadata: ToolSupport.pagination_metadata(meta)
-       }}
-    end
+    {:ok,
+     %{
+       test_cases:
+         Enum.map(test_cases, fn test_case ->
+           %{
+             id: test_case.id,
+             name: test_case.name,
+             module_name: test_case.module_name,
+             suite_name: test_case.suite_name,
+             is_flaky: test_case.is_flaky,
+             is_quarantined: test_case.is_quarantined,
+             last_status: to_string(test_case.last_status),
+             last_duration: test_case.last_duration,
+             last_ran_at: Formatter.iso8601(test_case.last_ran_at, naive: :utc),
+             avg_duration: test_case.avg_duration
+           }
+         end),
+       pagination_metadata: MCPTool.pagination_metadata(meta)
+     }}
   end
 
   defp build_filters(args) do
