+++ server/lib/tuist/mcp/components/tools/list_test_module_runs.ex
@@ -28,6 +28,7 @@ defmodule Tuist.MCP.Components.Tools.ListTestModuleRuns do
       "required" => ["test_run_id"]
     }
 
+  alias Tuist.MCP.Tool, as: MCPTool
   alias Tuist.Tests
 
   @impl EMCP.Tool
@@ -37,7 +38,7 @@ defmodule Tuist.MCP.Components.Tools.ListTestModuleRuns do
 
   def execute(conn, %{"test_run_id" => test_run_id} = args) do
     with {:ok, _run, _project} <-
-           ToolSupport.load_and_authorize(
+           MCPTool.load_and_authorize(
              Tests.get_test(test_run_id),
              conn.assigns,
              :read,
@@ -52,8 +53,8 @@ defmodule Tuist.MCP.Components.Tools.ListTestModuleRuns do
           status -> filters ++ [%{field: :status, op: :==, value: status}]
         end
 
-      page = ToolSupport.page(args)
-      page_size = ToolSupport.page_size(args)
+      page = MCPTool.page(args)
+      page_size = MCPTool.page_size(args)
 
       {modules, meta} =
         Tests.list_test_module_runs(%{
@@ -78,7 +79,7 @@ defmodule Tuist.MCP.Components.Tools.ListTestModuleRuns do
                avg_test_case_duration: mod.avg_test_case_duration
              }
            end),
-         pagination_metadata: ToolSupport.pagination_metadata(meta)
+         pagination_metadata: MCPTool.pagination_metadata(meta)
        }}
     end
   end
