+++ server/lib/tuist/mcp/components/tools/list_xcode_build_cache_tasks.ex
@@ -33,6 +33,7 @@ defmodule Tuist.MCP.Components.Tools.ListXcodeBuildCacheTasks do
     }
 
   alias Tuist.Builds
+  alias Tuist.MCP.Tool, as: MCPTool
 
   @impl EMCP.Tool
   def description,
@@ -43,7 +44,7 @@ defmodule Tuist.MCP.Components.Tools.ListXcodeBuildCacheTasks do
     build_run_id = Map.get(args, "build_run_id")
 
     with {:ok, _build, _project} <-
-           ToolSupport.load_and_authorize(
+           MCPTool.load_and_authorize(
              get_build(build_run_id),
              conn.assigns,
              :read,
@@ -60,8 +61,8 @@ defmodule Tuist.MCP.Components.Tools.ListXcodeBuildCacheTasks do
           end
         end)
 
-      page = ToolSupport.page(args)
-      page_size = ToolSupport.page_size(args)
+      page = MCPTool.page(args)
+      page_size = MCPTool.page_size(args)
 
       {tasks, meta} =
         Builds.list_cacheable_tasks(%{
@@ -86,7 +87,7 @@ defmodule Tuist.MCP.Components.Tools.ListXcodeBuildCacheTasks do
                cas_output_node_ids: task.cas_output_node_ids
              }
            end),
-         pagination_metadata: ToolSupport.pagination_metadata(meta)
+         pagination_metadata: MCPTool.pagination_metadata(meta)
        }}
     end
   end
