+++ server/lib/tuist/mcp/components/tools/list_xcode_builds.ex
@@ -5,6 +5,7 @@ defmodule Tuist.MCP.Components.Tools.ListXcodeBuilds do
 
   use Tuist.MCP.Tool,
     name: "list_xcode_builds",
+    authorize: [action: :read, category: :build],
     schema: %{
       "type" => "object",
       "properties" => %{
@@ -46,51 +47,49 @@ defmodule Tuist.MCP.Components.Tools.ListXcodeBuilds do
 
   alias Tuist.Builds
   alias Tuist.MCP.Formatter
+  alias Tuist.MCP.Tool, as: MCPTool
 
   @impl EMCP.Tool
   def description,
     do:
       "List Xcode build runs for a project. Only available for projects with build_system=xcode. The account_handle and project_handle can be extracted from a Tuist dashboard URL: #{Tuist.Environment.app_url()}/{account_handle}/{project_handle}."
 
-  def execute(conn, args) do
-    with {:ok, project} <-
-           ToolSupport.resolve_and_authorize_project(args, conn.assigns, :read, :build) do
-      page = ToolSupport.page(args)
-      page_size = ToolSupport.page_size(args)
-      filters = build_filters(project.id, args)
+  def execute(_conn, args, project) do
+    page = MCPTool.page(args)
+    page_size = MCPTool.page_size(args)
+    filters = build_filters(project.id, args)
 
-      {builds, meta} =
-        Builds.list_build_runs(%{
-          filters: filters,
-          order_by: [:inserted_at],
-          order_directions: [:desc],
-          page: page,
-          page_size: page_size
-        })
+    {builds, meta} =
+      Builds.list_build_runs(%{
+        filters: filters,
+        order_by: [:inserted_at],
+        order_directions: [:desc],
+        page: page,
+        page_size: page_size
+      })
 
-      {:ok,
-       %{
-         builds:
-           Enum.map(builds, fn build ->
-             %{
-               id: build.id,
-               duration: build.duration,
-               status: to_string(build.status),
-               category: if(build.category != "", do: build.category),
-               scheme: build.scheme,
-               configuration: build.configuration,
-               is_ci: build.is_ci,
-               git_branch: build.git_branch,
-               git_commit_sha: build.git_commit_sha,
-               cacheable_tasks_count: build.cacheable_tasks_count,
-               cacheable_task_local_hits_count: build.cacheable_task_local_hits_count,
-               cacheable_task_remote_hits_count: build.cacheable_task_remote_hits_count,
-               inserted_at: Formatter.iso8601(build.inserted_at, naive: :utc)
-             }
-           end),
-         pagination_metadata: ToolSupport.pagination_metadata(meta)
-       }}
-    end
+    {:ok,
+     %{
+       builds:
+         Enum.map(builds, fn build ->
+           %{
+             id: build.id,
+             duration: build.duration,
+             status: to_string(build.status),
+             category: if(build.category != "", do: build.category),
+             scheme: build.scheme,
+             configuration: build.configuration,
+             is_ci: build.is_ci,
+             git_branch: build.git_branch,
+             git_commit_sha: build.git_commit_sha,
+             cacheable_tasks_count: build.cacheable_tasks_count,
+             cacheable_task_local_hits_count: build.cacheable_task_local_hits_count,
+             cacheable_task_remote_hits_count: build.cacheable_task_remote_hits_count,
+             inserted_at: Formatter.iso8601(build.inserted_at, naive: :utc)
+           }
+         end),
+       pagination_metadata: MCPTool.pagination_metadata(meta)
+     }}
   end
 
   defp build_filters(project_id, args) do
